Optimization: Refresh only cells that have changed
Only cells that have changed, or could possibly change (dirty) cells are
being processed in each iteration of the game.

This means that not all the grid is being processed as before regardless
of whether the processed cells changed.

This is a massive optimization that saves a lot of cpu power